Move Autocompleter into a search folder
It needs its own space, because it's about to get refactored into
separate pieces.
